extracted add food and get all foods
