+++ lib/food.js
@@ -2,14 +2,16 @@ const API = 'https://qs-be.herokuapp.com/api/v1/'
 const LOCAL ='http://localhost:3000/api/v1/'
 const $ = require('jquery')
 const FoodRemove = require('./food-remove')
+const FoodAdd = require('./food-add')
+const FoodTable = require('./food-table')
 
 const Food = class Food {
 
   constructor () {
-    this.getAll()
+    FoodTable.getAll()
 
     $('.foods').on('click', '.remove', (event) => {
-      this.remove()
+      FoodRemove.remove()
     })
 
     $('.foods').on('click', '.food-cell', (event) => {
@@ -49,68 +51,10 @@ const Food = class Food {
       });
     })
 
-    $('#new-food').submit( (event) => this.addFood())
+    $('#new-food').submit( (event) => FoodAdd.addFood())
 
     $('form').on('submit', (event) => event.preventDefault());
   }
-
-  getAll() {
-    return $.ajax({
-    method: 'GET',
-    url: API + 'foods',
-    })
-    .then( (data) => {
-      this.makeFoodTable(data)
-    })
-    .fail( (error) => {
-      console.error(error)
-    });
-  };
-
-  remove() {
-    FoodRemove.remove()
-    // const button = event.target
-    // const id = $(button).data("food-id")
-    //
-    // return $.ajax({
-    //   url: API + `foods/${id}`,
-    //   method: 'DELETE',
-    // })
-    // .done(this.removeFromFoodTable(button))
-    // .fail( (error) => {
-    //   console.error(error)
-    // })
-  }
-
-  addFood() {
-    let form            = $('#new-food')
-    let newFood         = {}
-    newFood['name']     = $("input[name=food-name]").val();
-    newFood['calories'] = $("input[name=food-calories]").val();
-    $('.calorie-error').empty().remove()
-    $('.name-error').empty().remove()
-
-    if(newFood.name == "" && newFood.calories == "") {
-      this.caloriesError()
-      this.nameError()
-    } else if (newFood.calories == "") {
-      this.caloriesError()
-    } else if (newFood.name == "" ) {
-      this.nameError()
-    } else {
-      return $.ajax({
-        url: API + 'foods',
-        method: 'POST',
-        data: newFood
-      })
-      .then((data) => {
-        this.addToFoodTable(data)
-      })
-      .fail( (error) => {
-        console.error(error)
-      });
-    };
-  };
 
   editName() {
     const toEdit = {name: $(event.target).val(), calories: $(event.target).parent().next().text()}
@@ -150,25 +94,6 @@ const Food = class Food {
     });
   };
 
-  makeFoodTable(data) {
-    data.forEach( (food) => {
-      $('.foods tbody').prepend(`<tr data-food-id=${food.id}><td class='food-cell' >${food.name}</td><td class='calorie-cell'>${food.calories}</td><td><button data-food-id=${food.id} class='remove btn red darken-3'>-</button></td></tr>`)
-    })
-  }
-
-  // removeFromFoodTable(button) {
-  //   $(button).parent().parent().empty().remove()
-  // }
-
-  addToFoodTable(data) {
-    const food = data[0]
-    $(`<tr data-food-id=${food.id}><td class='food-cell'>${food.name}</td><td class='calorie-cell'>${food.calories}</td><td><button data-food-id=${food.id} class='remove btn red darken-3'>-</button></td></tr>`).prependTo('.foods tbody')
-    $('.calorie-error').empty().remove()
-    $('.name-error').empty().remove()
-    $('input[name=food-name]').val('')
-    $('input[name=food-calories]').val('')
-  }
-
   searchFoods(){
     let searchName = $('#food-filter').val()
     $('.foods tbody').html('')
@@ -183,15 +108,7 @@ const Food = class Food {
       console.error(error)
     })
   }
- caloriesError () {
-    $('.calorie-error').empty().remove()
-    $('#calorie-field').parent().after(`<div class='calorie-error'>Please enter a calorie amount</div>`)
-  }
-  nameError () {
-    $('.name-error').empty().remove()
-    $('#name-field').parent().after(`<div class='name-error'>Please enter a food name
-    </divl>`)
-  }
+
   updateCellName (data, tr) {
     const td = ($(tr).children()[0])
     $(td).children()[0].remove()
