+++ test/food-remove-test.js
@@ -1,0 +1,9 @@
+const assert = require('chai').assert
+const FoodRemove = require('../lib/food-remove')
+const $ = require('jquery')
+
+describe ( 'RemoveFood', () => {
+  it('is removes food from table',  (done) => {
+    
+  })
+})
