+++ test/index.js
@@ -1,0 +1,1 @@
+require('./food-remove-test')
